requirements.txt
uvicorn[standard]>=0.16.0
atom-ml[models]==4.9.1
fastapi>=0.70.1
pandas>=1.3.0
